Safari-609.1.11.1/LayoutTests/fast/media/window-name.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../resources/js-test-pre.js"></script>
</head>
<body>
<audio controls></audio>
<script>

description("Test that having an audio element in DOM doesn't magically change window.name.");

shouldBeEqualToString("window.name", "");

</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
