LayoutTests/fast/media/window-name-expected.txt

Test that having an audio element in DOM doesn't magically change window.name.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS window.name is ""
PASS successfullyParsed is true

TEST COMPLETE
